task_project/pyproject.toml
[tool.poetry]
name = "flask-web-app"
version = "0.1.0"
description = ""
authors = ["Suren Khorenyan <surenkhorenyan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-wtf = "^1.2.1"
flask-login = "^0.6.3"
psycopg = "^3.1.14"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
